+++ CHANGES.rst
@@ -8,6 +8,9 @@ Unreleased
 - Add Spanish translation
 - Add French translation
 - Drop Django 1.7 workaround from `select_subclasses()`
+- Drop support for `Django < 3.2`
+- Drop support for `Python 3.6`
+- Confirm support for `Django 4.1`
 
 4.2.0 (2021-10-11)
 ------------------